Introducing golibbpf
I decided that if we are going to do this, we need to do it right.

Right now we this is statically linking libbpf into the binary (very quickly I might add)
But we need to figure out how we want to do dependency management here. I hate to say it
but I think gitmodules might be the best option...

Anyway - scaffolding for golibbpf
It compiles and runs if you have libbpf headers on your system

In arch linux: pacman -S libbpf

Signed-off-by: Kris Nova <kris@nivenly.com>